+++ test/test_display_logic/test_main.cpp
@@ -24,6 +24,30 @@ void test_persisted_setting_ranges() {
   TEST_ASSERT_FALSE(sdd::isValidCityCode(102000000));
 }
 
+void test_weather_value_ranges() {
+  TEST_ASSERT_TRUE(sdd::isValidTemperature(-80));
+  TEST_ASSERT_TRUE(sdd::isValidTemperature(80));
+  TEST_ASSERT_FALSE(sdd::isValidTemperature(-81));
+  TEST_ASSERT_FALSE(sdd::isValidTemperature(81));
+  TEST_ASSERT_TRUE(sdd::isValidHumidity(0));
+  TEST_ASSERT_TRUE(sdd::isValidHumidity(100));
+  TEST_ASSERT_FALSE(sdd::isValidHumidity(101));
+  TEST_ASSERT_TRUE(sdd::isValidWeatherCode(0));
+  TEST_ASSERT_TRUE(sdd::isValidWeatherCode(999));
+  TEST_ASSERT_FALSE(sdd::isValidWeatherCode(1000));
+}
+
+void test_iso_date_validation() {
+  TEST_ASSERT_TRUE(sdd::isValidIsoDate("2024-02-29"));
+  TEST_ASSERT_TRUE(sdd::isValidIsoDate("2026-08-03"));
+  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2025-02-29"));
+  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2026-13-01"));
+  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2026-04-31"));
+  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2026-8-03"));
+  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2026-08-03T00:00:00"));
+  TEST_ASSERT_FALSE(sdd::isValidIsoDate(nullptr));
+}
+
 void test_bar_widths_are_clamped() {
   TEST_ASSERT_EQUAL_UINT8(0, sdd::temperatureBarWidth(-20));
   TEST_ASSERT_EQUAL_UINT8(10, sdd::temperatureBarWidth(0));
@@ -44,13 +68,52 @@ void test_aqi_boundaries_follow_the_chinese_index() {
   TEST_ASSERT_EQUAL(sdd::AqiLevel::Severe, sdd::classifyAqi(301));
 }
 
-extern "C" void setup() {
+struct TestBannerText {
+  size_t textLength;
+  size_t length() const { return textLength; }
+};
+
+void test_carousel_skips_empty_pages_and_wraps() {
+  const TestBannerText pages[] = {{3}, {0}, {0}, {5}, {0}};
+  TEST_ASSERT_EQUAL_INT(0, sdd::nextNonEmptyIndex(pages, 5, 0));
+  TEST_ASSERT_EQUAL_INT(3, sdd::nextNonEmptyIndex(pages, 5, 1));
+  TEST_ASSERT_EQUAL_INT(3, sdd::nextNonEmptyIndex(pages, 5, 3));
+  TEST_ASSERT_EQUAL_INT(0, sdd::nextNonEmptyIndex(pages, 5, 4));
+  TEST_ASSERT_EQUAL_INT(0, sdd::nextNonEmptyIndex(pages, 5, 10));
+}
+
+void test_carousel_handles_no_available_pages() {
+  const TestBannerText emptyPages[] = {{0}, {0}, {0}};
+  TEST_ASSERT_EQUAL_INT(-1, sdd::nextNonEmptyIndex(emptyPages, 3, 0));
+  TEST_ASSERT_EQUAL_INT(-1, sdd::nextNonEmptyIndex(emptyPages, 3, 2));
+  TEST_ASSERT_EQUAL_INT(-1, sdd::nextNonEmptyIndex<TestBannerText>(nullptr, 0, 0));
+}
+
+void test_long_banner_scroll_offsets_are_bounded() {
+  TEST_ASSERT_EQUAL_INT(0, sdd::bannerMaximumOffset(150, 150));
+  TEST_ASSERT_EQUAL_INT(38, sdd::bannerMaximumOffset(188, 150));
+  TEST_ASSERT_EQUAL_INT(25, sdd::nextBannerOffset(0, 38, 25));
+  TEST_ASSERT_EQUAL_INT(38, sdd::nextBannerOffset(25, 38, 25));
+  TEST_ASSERT_EQUAL_INT(38, sdd::nextBannerOffset(38, 38, 25));
+}
+
+int run_display_logic_tests() {
   UNITY_BEGIN();
   RUN_TEST(test_brightness_validation_and_pwm_mapping);
   RUN_TEST(test_persisted_setting_ranges);
+  RUN_TEST(test_weather_value_ranges);
+  RUN_TEST(test_iso_date_validation);
   RUN_TEST(test_bar_widths_are_clamped);
   RUN_TEST(test_aqi_boundaries_follow_the_chinese_index);
-  UNITY_END();
+  RUN_TEST(test_carousel_skips_empty_pages_and_wraps);
+  RUN_TEST(test_carousel_handles_no_available_pages);
+  RUN_TEST(test_long_banner_scroll_offsets_are_bounded);
+  return UNITY_END();
 }
 
+#ifdef ARDUINO
+extern "C" void setup() { (void)run_display_logic_tests(); }
 extern "C" void loop() {}
+#else
+int main() { return run_display_logic_tests(); }
+#endif